Improved Key Assumptions
Improves key assumptions by ignoring already checked locations that are not keys. For example: If a hint says Forest Floormaster isn't a key and you check it as junk, the hover boots requirement isn't needed for key assumptions. Same principle for checking Water Central Pillar and Zora Tunic. Probably overkill for BotW and Ganons but w/e. Probably also improvable code-wise :] Also not 100% sure it's correct which is pretty iffy with the tournament going on lol.